requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
dataclasses-json==0.5.14
datasketch==1.6.3
datasketches==4.1.0
debugpy==1.8.0
decorator==5.1.1
duckdb==0.8.1
emoji==2.8.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.99.1
fastjsonschema==2.18.0
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kaleido==0.2.1
kiwisolver==1.4.5
lambdaprompt==0.5.5
langchain==0.0.334
langchain-experimental==0.0.40
langsmith==0.0.63
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbformat==5.9.2
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.6
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.10
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandas-llm==0.0.6
pandasai==1.2.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.17.0
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.1
pydantic==1.10.12
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-environ==0.4.54
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
RestrictedPython==6.2
rich==13.6.0
rpds-py==0.10.3
scipy==1.11.2
six==1.16.0
sketch==0.4.2
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
stack-data==0.6.2
starlette==0.27.0
streamlit==1.28.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tokenize-rt==5.2.0
tokenizers==0.14.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
unstructured==0.10.14
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
